Functions to get all org files in Roam DB tagged for export.
Rather then a brute force text search, I was hoping to do this with
the Roam sqlite db, but that appears to only be for caching. Parsing
the top ten or so lines for a couple thousand files also doesn't take
as long as I thought it would so the point is moot.