Replace publish-record.csv with a link list we create from blog-info
This approach strikes me as slap-dash, but it doesn't require us to
maintain a separate file to handle org-roam id links so I'll run with
it for now.

Ehhhh, I could use a dynamic var, an atom, or maybe just attach the
link-list to every piece of blog-metadata.

This is fine for now.